Drop subappendix type
This reference type had been introduced as half-measure for some sorting
problems for the appendix which were later on properly fixed, and I forgot to
remove it then.  And its existence represents an inconsistency vis-à-vis
sectioning commands, as shown in the discussion on issue #1 (https://github.com/gusbrs/zref-clever/issues/1#issuecomment-981128721).